requirements.txt
openai
jinja2
python-dotenv
pytest
streamlit